refactor: markdown processor
